KokkosBatched - trsm host
